Always print max degree and culprit constraint
With this commit the (table, constraint) tuple that determines the max
degree is always printed to standard out when the constructor of the
STARK object is run. This will make optimizations easier as you can
always see which constraint is the constraining factor for a shorter FRI
domain.

This closes #25.