Use 'path' as dependency fallback in Cargo.toml
Instead of switching between crates.io and a local copy of twenty-first
by commenting in another dependency, specify both `version` and `path`.

This corresponds to having "multiple locations" as described in The
Cargo Book:

https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#multiple-locations

> It is possible to specify both a registry version and a `git` or
> `path` location. The git or path dependency will be used locally (in
> which case the `version` is checked against the local copy), and when
> published to a registry like crates.io, it will use the registry
> version. Other combinations are not allowed.

Since both the repositories `triton-vm` and `twenty-first` contain a
crate in a sub-directory, one can create a symlink in the right position
to make the relative path `../twenty-first` correspond to the crate from
the local clone on disk:

```
~/Projects/triton-vm/triton-vm $ cd ..
~/Projects/triton-vm $ ln -s ../twenty-first/twenty-first twenty-first
```

When the directory is not present, the crate on crates.io is used, and
when present and correctly positioned, the crate on disk is used. One
can either clone directly from

https://github.com/Neptune-Crypto/twenty-first

or clone from a remote fork of that repository.