Remove TableId from BaseTable
This field was added in order to be able to iterate a table collection
and create a HashMap mapping a unique table identifier to some memoized
value, e.g. a vector of constraints.

Instead, vectors of constraints are stored on the BaseTable object.

Having TableId on the BaseTable overlaps slightly with putting BaseTable
inside of concrete table structs, like JumpStackTable.