Add function to apply known challenges to circuit constraints
This can be called prior to evaluation to speed it up.

To my surprise, though, applying the challenges followed by constant
folding does not reduce the number of nodes in any circuits. This means
that *all* our challenges are multiplied or subtracted/added to an
input, no challenges are combined with constants or with each other in
any constraints.