Add new leaf type to circuits: B field element
Some operations can be done faster if we know that the leaf constant is
a B field element. E.g. multiplying a B field element with an X field
element is three B field multiplications, whereas it's 9 for two X field
elements.

So when we autogenerate code, we can generate faster code if we know
that the leaf is a B field element.