Merge branch 'thv/circuits'
Replace multivariate polynomial description of transition constraints
with a circuit-based one.

This introduces a monad-like structure that gives each node in the
multitree/circuit/DAG a unique ID *and* the monad checks if this
specific node has been constructed before and reuses that if it has.
Once the multitree is constructed, the circuits can be wrapped out of
their monad through the function `consume`. Within the monad leafs can
be constructed by the `ConstraintCircuitBuilder` and nodes can be built
from other nodes through the infix operators of `Add`, `Sub`, and `Mul`.

Currently the circuits are flattened immediately after being constructed
through the function `partial_evaluation`. A speedup will not be
observed before this is changed.

Future improvement suggestions:
- Remove the leaf type for circuits that is Randomized Input and replace
it with the composition of `challenge * Input`. This will improve the
circuit's reuse of nodes and should thus help us get rid of a few more
terms in the final calculations.
- in the circuit's binop method, when a new node `a + b` is constructed,
  check if `b + a` is already in the multitree. Do the same for
multiplication.
- Implement a function that eliminates the abstract challenges from the
  circuits and replaces them with concrete X field values. This
  simplified multitree can then be constant folded again for further
  simplification.
- Implement an `eval` function so the circuits can be used for
  evaluation.