Fix inconsistent use of challenges in trace/constraint test
Prior to this commit, this test used to different sets of challenges for
constructing the extension constraints and then later in their
evaluation. Since this argument was until now not used in the evaluation
(since the challenges were contained in the multivariate polynomials
describing the constraints), this problem did not surface. But now that
we're autogenerating code to evalaute the constraint polynomials, this
problem surfaced and it just cost me almost half a day's work.