Change the `Input` type of circuits to allow handling of single row constraints
This commit allows single row constraints to be expressed as a circuit
and to work with the compiler for auto-generating efficient constraint
evaluating Rust code.

This is achieved by changing the input type from holding a `usize` to
holding a `InputIndicator` type that can hold either single row and dual
row constraints. In the generated code the variable name of an input
type is assumed to be the same as the `Display` implementation of the
`InputIndicator` types.