Add option of printing profile output as-you-go
By setting the environment variable `PROFILE_AS_YOU_GO` you can now get
the names from the profiler printed as the program runs. This will not
be indented nicely like the final report, but this commit does not touch
how the final report is printed, so you still get that.

This is introduced because it's nice to have for very long programs,
some of our benchmarks/tests still take hours on the fastest computers
and it's nice to be able to see during the run how far it has come.