triton-opcodes: Add support for 'call' bracket syntax
`[foo]` is syntax sugar for `call foo`.

This is supposed to give the feel that `foo` is an instruction without
naively allowing for e.g. `skiz foo`, since that would require for a
`skiz call foo` (or now `skiz [foo]`) in the most likely event that
`foo` maps to more than one instruction.