ensure `padded_height` is never overshadowed by `num_trace_randomizers`
Concretely, `padded_height` is at least as big as
`num_trace_randomizers` rounded up to the next power of two. This
guarantees a unique `padded_height` given an `interpolant_degree` and
the `num_trace_randomizers`. This, in turn, allows to determine the
`padded_height` given the FRI domain size.

Note: for the default STARK parameters, this does not increase the
`padded_height` under any circumstances.