Fix opcode error and add regression test
`lt` decreases the opstack size, and this must be reflected in its
opcode. This bug was discovered by attempting to prove/verify all
`tasm-lib` snippets.

New version of opcodes must be used for this fix to work. That
dependency upgrade is not included in this commit, but the fix to
`triton-opcodes` is.

Co-authored-by: Jan Ferdinand Sauer <ferdinand@neptune.cash>