fix: Two bugs in constant-folding PBT
First bug relates to stack overflow and is addressed by letting circuit
monads go out of scope when they are no longer needed. Previously, we
made deep copies of the circuit monads and kept them all in scope within
the entire test body.

Second bug was much more complicated and related to node substitution
in a erroneous setup where different circuit-builders were used. The
`deep_copy` function creates a new builder, and circuits with different
builders where used to construct binary expression. This leads to a risk
of collission of IDs which will make constant-folding misbehave. The
solution is to rewrite the test such that binary expressions are always
formed from nodes which share the same circuit-builder.

One more change is prudent: The circuit builder should carry around a
random-generated ID to ensure that no one forms binary expressions from
two nodes that do not share the same builder. I plan to address that in
the next commit.