fix!: don't treat randomizer polynomial special
Non-critical bug: prover & verifier used to suggest that the randomizer
polynomial should be treated in a special way for the DEEP update.
Now, they do not suggest this anymore. No behavior changes.

An alternative approach is to add the randomizer polynomial into the
combination codeword without multiplying it by a random weight first.
This approach is not taken for now due to upcoming changes to the way
the random linear sum's weights are constructed.

BREAKING CHANGE: Delete method `fri_domain_randomizer_polynomials`,
which always returns an empty vector, making it de-facto useless.