refactor: provide API to overwrite cache decision
Allow forcing the LDE trace caching behavior even if setting an
environment variable is infeasible. 

Rename the environment variable to `TVM_LDE_TRACE`, accepting any of the
values `cache` and `no_cache` where capitalization does not matter.

Previously, the decision whether to allocate memory for the low-degree
extended trace table was made ahead of time, and the actual allocation
happened later. Now, the allocation is attempted, and if it does not
succeed, the just-in-time codepath is chosen.

Additionally, performance has improved: `sysinfo`'s function how much
`.free_memory()` is available is apparently prohibitively slow to
execute: ~25ms on a fast system and ~40ms (!) on a standard laptop.