refactor(profiler)!: make `TritonProfiler` private
Change the interface for using the internal profiler. Most notably,
at most one instance should exist at any given time and per thread.
The already-existing macros like `prof_start!` and `prof_stop!` work
as before, and use the one profiler instance.

This design is less intrusive, as it does not require passing the
profiler as an argument to every function that can or should be
profiled.