Refactor: factor out population of CJD SLD column
Before, all extension columns are populated through a single loop over
all rows. The objective is to factor out all columns and populate them
in separate functions. Then we can parallelize over all (function,
extension-column) pairs to make use of more cores as we populate them.

This commit is the first step in this rather large refactor; it
factors out the population of just one extension column. The others
still have to be factored out.

Co-authored-by: Thorkil Værge <thor@neptune.cash>