perf!: Halve number of combination codeword checks
In line with #285 we only need to verify the combination codeword in
`num_collinearity_checks` many points and not in 2 times this number.
This is because we we inside of FRI explicitly commit to the FRI
codeword in `2 * num_collinearity_checks` many points.

This also reduces the number of trace randomizers needed.

Co-authored-by: Ferdinand Sauer <ferdinand@neptune.cash>