feat: Introduce instruction `recurse_or_return`
Instruction `recurse_or_return` behaves – surprise! – either like
instruction `recurse` or like instruction `return`. The (deterministic)
decision which behavior to exhibit is made at runtime and depends on the
current stack as well as the instruction's immediate argument `i`.

Argument `i` indicates one of the stack's top 16 elements. Let the value
of the indicated stack element be `a`, and the value of its successor
element `b`. In the slightly special case of `i` == 15, the indicated
element's successor is taken to be stack element with index 0.
If `a` != `b`, then `recurse_or_return` acts like instruction `recurse`,
else like `return`.