fix: Ensure node index for `merkle_step` is u32
Previously, a dishonest prover could have modified the node index used
in instruction `merkle_step` to be a non-u32. This could have led to
proving inclusion of a certain leaf in a tree of a (possible much)
greater height than assumed by the verifier. While this is not
exploitable under the assumption that the hash function Tip5 is
collision resistant, such an argument is subtle.
The fix as implemented in this commit removes the subtlety by forcing
both, the input node index and the produced parent node index to be u32.