feat: Introduce instruction `merkle_step_mem`
Similar to existing instruction `merkle_step`, the new instruction
allows verifying a Merkle authentication path efficiently. Unlike the
existing instruction, that authentication path lives in Triton VM's RAM.
This allows to verifiably use the _same_ authentication path multiple
times. This is relevant when updating a leaf in a Merkle tree in a
verifiably correct manner.