refactor(test): Simplify spec generation
If the constraints are not in place, the corresponding error message is
well suited to figure out what is going on. Since this is a test, not a
piece of production code, that's all that's needed. Adding special
error handling logic, especially with the to-be-avoided “catch_unwind”,
is overkill.

changelog: ignore