test: Catch changes in constraint evaluators
This commit adds a test to verify that the automatically generated code
for evaluating the constraints has not changed between (a) writing or
updating the test; and (b) running the test. This test spans a large
scope and therefore can catch a large class of errors, ranging from
non-determinism to unintentional changes in the constraint circuit.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>