fix: Hunt, catch & squash weird shrinking bug
The superficial issue was that degenerate arbitrary instances
unreachable! code. This is now caught and this catching causes the
proptest to pass.

The deeper issue was that the first unreachable case triggers the
proptest to start shrinking and then all hell breaks loose. We observe
that the shrinker somehow starts emitting *more complex*
multicircuits.

Co-authored-by: Thorkil Værge <thor@neptune.cash>