ci: Skip degree-lowering benchmark in nextest CI run
This is achieved by splitting the repository-wide `nextest` command
into one `nextest` command for each crate. For crate
`triton-constraint-builder` the `--all-targets` flag is dropped.

Also, the flag `--all-targets` is dropped from the coverage data
collector. As a consequence, benchmarks are not treated as tests
for the purpose of computing code coverage. And in particular, the
degree lowering benchmark is skipped also.