style: Disambiguate with new types
Function `apply_substitution` used to take three `usize`s (among other arguments)
for different purposes. This is considered bad design because it is easy to mess
up the order of the arguments leading to runtime failures or erroneous behavior.
Declaring and using the type aliases leverages the compiler's type system to
combat misuse.