perf(RAM): Drop cached polynomials
The cached polynomials represent the same information as the original
randomized trace, except in a different basis. If they are needed, the
relevant randomized trace column is interpolated on the fly.

Through various techniques, like batched barycentric evaluation, the
drawback of losing the cached polynomials is offset almost completely.
In the process, the code becomes notably simpler and, more importantly,
requires considerably less RAM.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>