perf(RAM): Clear unusable caches
It is possible that the quotient domain codewords are cached for the
main table, but not for the auxiliary table. To compute the quotients,
either both or neither are needed.[^1] For peak memory consumption, it
is beneficial to clear any unused cache.

Throwing away any cache here incurs a performance penalty later, when
revealing the rows indicated by FRI. This is deemed an acceptable
tradeoff when peak memory usage is a larger issue than compute time.

[^1]: Code using exactly one cache could exist, but oh! the engineering.