fix: Improve error message on internal error
When manually initializing a Triton VM state with a stack that is too
shallow, Triton VM execution will fail. Previously, the error message
presented to the user was `attempt to subtract with overflow` if debug
symbols were enabled, and `index out of bounds` in release mode. Now,
the error message is more meaningful.