refactor(verifier): Remove parallelism
A downstream consumer of Triton VM uses the verifier in an asynchronous
context while running its own `rayon` threadpool. This, in combination
with Triton VM's use of `rayon`, is suspected to cause odd behavior.

Remove parallelism through explicit use of `rayon` from the Triton VM
verifier, with the option to re-introduce it when the underlying issues
are better understood.