WIP: add debug statements to identify high-security STARK problems
Thorkil raises the point that given Tip5 and XFieldElement's extension
degree 3, it doesn't even make sense to have a security level this
high. Check that out, it's probably true. Once the boundary by these
two things are identified, an update to the documentation plus an
assert statement should be enough to fix the problem.