feat(profiler): Report memory consumption
Track the resident set size (rss) whenever a task is started or stopped,
and include the maximum encountered rss in the final report.
Additionally, for each task, report the (averaged over number of
iterations) rss growth or shrinkage. This helps to understand _where_
the RAM usage comes from, not just how big it is over all.

With the current configuration, it is possible that the maximum rss is
underreported. Also, some systems or system configurations might make
querying for the rss impossible, in which case placeholders communicate
the lack of information.

Also, reword most documentation of the profiler to be less specific to
Triton VM since it's a pretty generic profiler, really.