fix: Use rayon's thread count for parallelization
Previously, `std::thread::available_parallelism()` was used directly to
determine the size of the workload to be used in individual parallel
workers. The number of available parallelism is difficult (usually
impossible) to configure by users.

Now, Triton VM uses `rayon::current_num_threads()` instead. This means
that implicitly, rayon's environment variable `RAYON_NUM_THREADS` is
taken into account when distributing work to parallel workers. This can
be especially useful on machines with a large number of cores, as it can
lead to reduced memory requirements of Triton VM when the RAM-frugal
execution path is chosen.