docs: Fix incorrect transition constraint
In the specification for the lookup table, the update rule for the
transition constraint of the public evaluation argument was written as a
permutation argument. Since the argument is an evaluation argument, the
update rule corresponding to evaluation arguments is used.

Note that is a change to the documentation only; the code does not
contain this bug.